Add MVSG-HV documentation
MVSG-HV has been in Xyce since release 6.11, but never added to the
reference guide (the corresponding issue was left open because the
intent was to get more evaluation of the model before declaring
victory on it).

Rather than leave that issue open forever, I'm trying to close it out.
I've added tables of instance and model parameters as well as output
variables, and a link to the NEEDS/nanohub pages that document the
model.  I could find no direct link to any official page about the
model other than the Nanohub one.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=901