Add two new DDX tests that fail
I found a use case that I had not accounted for in the DDX operator.

If I have this expression:  DDX(F1(P1),P1) then the ddx operator doesn't
successfully resolve P1, and thus thinks it can't evaluate DDX.

Currently, DDX just searches the lists of parameters, voltage nodes and
device currents it has lists of, and somehow, the argument to F1,
whether it be a parameter or a voltage isn't in the list.

But I think it would be in the list for this expression:  F1(P1).  So
this is solvable.