Fix for ddx with .funcs
I found a use case for ddx that I had not accounted for.    I explained
this in a previous commit, but basically the new expression library was
busted if the expression applied the DDX operator to a function call,
especially if the derivative variable was one of the arguments to the
function call.

So, in other words:

.PARAM P1=5.0
.FUNC F1(X) {X+1}
{DDX(F1(P1),P1)}

did not work.  It gave an "unresolved" error.

The fix was to (1) do better book-keeping in setting up the DDX
operator, and (2) move most of that book-keeping from the ddxOp
constructor to the ddxOp::val() operator.  At the time of
construction, functions are not 100% resolved, so even with better
book-keeping, the derivative entity can't be found.  But, at the
time of evaluation (when val() would be called), everything
should be in order.

It works now, and I've created two unit tests that prove it.