Implement the "isTimeDependent" function
I had, for some reason, left this function unfinished, and it was
unconditionally returning true.  This probably slowed the code
down a bit.

If a device parameter is dependent on an expression, and this function
returns "true" it causes the relevent device(s) to re-evaluate this
parameter at every time step.    That mostly won't cause the wrong
answer (although in one corner case, numerical sensitivities, it did),
but it forces the code to waste resources.

I discovered this when I finally dug into why the SENS/bug1070sonPart2
test was failing.  This was a simple RC tranisent sensitivity test, in
which two versions of the same circuit were compared, one that used
global parameters and one that did not.  They should have produced
matching answers, and they did not.

My guess is that the numerical sensitivies in Xyce don't work at the
moment when attempting to compute a sensitivity w.r.t. a time dependent
(or solution dependent) parameter.    In this case, the code was
mistakenly thinking this was the case and that was causing it to fail.