Fix parmetis usage
In commit e3bd9b19, ParMETIS was made an optional component of Xyce,
because its copyright precluded it being a mandatory component under
some conditions.  That was done by adding a "--enable-parmetis" option
to configure and a new macro Xyce_USE_PARMETIS that was enabled if
that option was chosen.

That was a bad approach, because Xyce does not itself use ParMETIS
directly --- it's used only by Trilinos, and if we want to support it
we have got to have it compiled into Trilinos in the first place.
Allowing it to be a manual selection is wrong because if Trilinos is
NOT built with ParMETIS and Xyce this option is selected, then it
won't work, and if Trilinos *IS* built with ParMETIS and the user
doesn't select "--enable-parmetis" then configure won't probe for the
libraries and there will be link errors at the end of the build.

Therefore, very shortly after commit e3bd9b19, I made commit 53c8d039,
which checks if EpetraExt_config.h defines HAVE_LIBPARMETIS,
indicating that Trilinos was compiled with ParMETIS support.  If so,
it looks for parmetis libraries, adding them to the link line if found
and aborting if not.  It does not set ANY macros either way.

Unfortunately, I neglected the single ifdef in N_LAS_TransformTool.C
that actually used the "Xyce_USE_PARMETIS" --- since nobody was
setting this anymore, Xyce would gleefully link in ParMETIS as needed
to satisfy Trilinos, but disable the code that actually allowed a user
to *select* the partitioner it should provide.

The correct ifndef should be HAVE_LIBPARMETIS, which is provided by
the various Trilinos headers included in that file if in fact Trilnos
is built using it.

We stopped building and testing Xyce with ParMETIS as soon as it
became optional, and so this never came up until a user tried to build
with it.