Better support for bsrc currents
Bsrc currents are a little tricky in Xyce, because they can either be a
lead current (for I={exp} devices) or a solution current (for V={exp}
devices).

If an expression depends on a Bsrc current, it is impossible to know at
expression parsing time what kind of current it is.  Sometimes this
matters, and sometimes this does not.

Here are two cases where it matters.  There may be others.

(1) if it is a lead current, then the device package has to be told
about it, during setup.  If it isn't told about it, then it won't
compute or store it.  Lead currents aren't necessary for the solve, so
they should only be computed if they've been requested for output.

(2) If the expression is used in another Bsrc device, it cannot depend
on a lead current.  It can only depend on solution variable currents.
If you try to use a Bsrc that depends on a lead current it should fail
with an error during the setup.

When parsing an expression (like {2.0*I(B1)}), it is easy to determine
that "B1" is a Bsrc, as it starts with "B".   But not possible to know
if it is a lead or solution current.

I had been unconditionally making I(B*)  devices assume they were
solution variable currents.  This worked fine most of the time, except
for when I(B*) was used on the .print line, and I(B*) was a lead
current, and the particular I(B*) wasn't being requested otherwise.
In that case, the lead wasn't requested from the device package, so it
couldn't be resolved later.

The reason is this.  The old expression library used the function
get_names(XEXP_LEAD,leads) to get a list of lead current names.  In the new
library, I have a new function getLeads(leads).  They theoretically do the
same thing.  When I put I(B*) unconditionally in my list of solution
currents, it was never included in the vector of strings returned by
this function.

But, if I unconditionally put it into my list of lead currents, this
casued other problems.  It meant that the solution currents were not
found properly.

So, I created a new container and accessor function, just for Bsrc
currents.  These are included in what gets returned by getLeadCurrents.
This seems to have fixed the issue.  At least one regression test had
been failing and now it is not.