Improve Handling of WHEN Time for DERIV-WHEN Measure
This commit fixes the handling of the WHEN time for the DERIV-WHEN
measure, for AC, DC and NOISE measure modes, if both FROM and TO are
specified.  It covers the "corner case" of the FROM and TO values
being completely within a sweep interval.  The correct interpolated
WHEN time is now found for that case.  Previously, a measure for
that corner case would be reported as "FAILED".

SON Bug 1304
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1304