Removes calls to addInternalNode in loadNodeSymbols
There can be 4 or 5 external nodes for this device, but since
they are external nodes they do not need to be added to the
symbol table in loadNodeSymbols().  This method expects to
add node symbols for nodes that are local (or owned by) the
processor.  Any external nodes can be owned by another processor
in parallel, as is the case when Xyce hangs during the
loadNodeSymbols() method in the Device::PowerGridTransformer
class.  The N or Phi node is actually owned by another processor,
because the voltage node is owned by another processor.  This
is a bad device distribution, but regardless, it is not necessary
to load symbols for N or Phi.  Removing the calls to
addInternalNode() in the loadNodeSymbols() method eliminates
the failed parallel execution of these tests.

This addresses SON bug 1311