Adds Util::UParamList, an unordered_set parameter list implementation
The ParamList object is really a std::list, or linked list, object.
That means that searches through it are linear.  If many searches need
to be performed this can be costly.  Furthermore, to search through
this object string comparisons are made.  This only adds to the cost
of this object.  It has been identified as a primary cause for excessive
setup times in Xyce's netlist parsing in issue #35.

This commit adds an alternative container for parameter lists, which
is an unordered_set.  Since the ParamList should contain Param objects
with unique tags, a set makes more sense than a linked list.

Issue #35