Utilizes new Util::UParamList for resolvedParams_/unresolvedParams_ objects
The Util::ParamList object, which is a std::list, is inefficient for PDKs
that have many parameters that need to be resolved.  The list of resolvedParams_
within any part of the circuit hierarchy should be unique, so a std::set
object would be more fitting of the design.  Using an unordered_set, such
as the new UParamList, will allow for more efficient searching that will
use string hashes instead of string comparisons.  This should also improve
inefficiency of the object, and the resulting netlist parsing and setup
phase of Xyce.

This addresses issue #35