Improve .print expr error reports, esp power
One of the regression tests that has been failing is
POWER/power_unsupported_devices.cir.  The reason it was failing was, in
some ways, very small.   At first glance it didn't seem like a
meaningful discrepancy, but as I dug into it, it seemed a little
more important.

This regression test attempts to get power outputs on the .print line
that are invalid, and then checks to see that the correct error messages
are reports.  One of the fields on the .print line is an expression,
but all the rest are not.   The non-expression errors got reported
correctly with the new expression library (not surprising, as it wasn't
involved) but the expression error simply did not get reported.

The reason for this was that I had set up the outputs "group" for the
new expression library to not set things up until it was actually time
to .print.  (ie, after at least one solve had completed).  This was
enough later that the code exited before that error could be reported.

I've updated the code so that the setup happens earlier and now the
error is gone.  Also, I had to modify the powerOp class so that it knows
if the power was requested via the "P()" operator, or the "W()"
operator, so that the error message knows which one to use.

This commit also includes the addition of the standard Xyce file headers
to most (hopefully all) of the new expression source files.  I'd been putting
off adding them so I am glad that I finally did it.

Also, while doing it, I noticed a number of files in the repo that I no
longer use.  These are all files related to the original test programs
that I wrote.     That test program pre-dated my use of unit tests.  It
was also a program that I mainly used when I was developing the new
expression library entirely outside of Xyce, in a separate non-Xyce
repository.

When I migrated all this work to be integrated into Xyce, I copied over
more files than I needed.  So, the files that are being removed are ones
that should never have been copied.