Add meaningful breakpoints test for stpOp & fix
I have been running the unit tests thru lcov again, and have found a
number of oversights in my unit testing.

One was related to my work a few weeks ago on breakpoints for
comparision operators, as well as the STP operator.  This was the type
of breakpoint that is computed via a Newton loop.

With this commit I have added two tests of this breakpoint detection
code, for the STP operator.  I still have to add more, for other
comparison operators but that will come later.

While doing this, as usual, I did find some small mistakes in the
expression library and have fixed them with this commit.

Also, included is some cleanup of the newExpression.C file.