Cleanup, plus slightly better error reports
This is mostly a cleanup commit.  But I also made a small change to
error reporting, which usually prints the expression string.

For the new expression library, the lexer & parser is
case-insensitive, but it wasn't necessarily case-insensitive
when interacting with the rest of Xyce.

So, I always do a "toUpper" on the expression string.  I had been using
the up-cased expression string for error reporting.

For error reporting, and for consistency with the old expression
library, it is better to use the original, unmodified expression string.
So, with this commit I am doing that now.

I also got rid of the N_UTL_Expression::set function, which was
worthless, and that I didn't really trust at this point.