Add missing call to popActiveAnalysis()
In analyses like MPDE and HB,  you frequently see the following code
pattern:

  // allocate analysis (transient) object
  analysisManager_.pushActiveAnalysis(&transient);
  // other details
  returnValue = transient.run();
  // other details
  analysisManager_.popActiveAnalysis();

This happens b/c there is a top level analysis and a subordinate one,
and the subordinate once changes one or more times.

The MPDE manager was missing one call to popActiveAnalysis().  This
was  *mostly* not a problem, but it meant that the analysis object stack
contained an invalid pointer after the analysis "transient" went out of
scope, which it did soon after "run" was called on it.
This invalid pointer was temporarily at the "back" of the analysis
object stack, which was a problem.  It was eventually no longer at
the back, as there later would be another call to pushActiveAnalysis,
thus putting a valid pointer at the "back".

It was bad that a NULL pointer was there at all, but most of the
time it was in a benign location, as it wasn't in the "back" position.

Some of my recent work on the new expression library exposed this
problem.  I added a call to "evaluateFunction" in the
ExpressionData::setup function, which in the MPDE case is called
*before* one of the "pushActiveAnalysis" calls.  So, it attempted
to access data thru the NULL pointer.

This commit fixes the problem.