More breakpnt tests & bugfixes; table/spice src's
Added 2 new unit tests, to check breakpoints from
time-dependent table sources. One with the table invoked directly, and
the other thru a .func.  Initially, the first test passed, and the
second test (thru the .func) failed.

Fixing this was a matter of bookeeping.  Most of my recent bugfixes have
been very similar.  The parser automatically compiles lists (actually
std::vector's) of different types of operators, that need to be treated
in particular ways.  This includes time dependent sources such as
spice_pulse, and table, etc.

At the end of the parse phase, these containers are incomplete for
expressions that have external dependencies such as via .func. So,
extra work has to be done to take care of the bookeeping for externally
dependent expressions.  For a number of these types of objects, I
neglected to implement this extra bookeeping.  This (time dependent
sources) is the most recent example.

This is fixed now, so correct breakpoints are returned from tableOps
that are inside of .funcs.