Begin to add port_connected support
This commit is just barely enough to allow port_connected to be
recognized and the ADMS data structure to be expanded to keep track of
optional nodes.

A huge amount of additional work will be required to actually make
port_connected work properly.  But this was the first thing that
needed to be out of the way.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=390
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/24