Begin work to support port_connected
This commit is a tiny step in the direction of supporting
$port_connected in verilog-a.

It sets up a vector, "portsConnected_" of bools equal in size to the
number of declared ports in the module (that is, what ADMS considers
"external" nodes).  It initializes them all to "true", then checks how
many nodes were given on the instance line.  If it is fewer than "all
of them", it sets the elements of portsConnected associated with the
unspecified nodes to "false".

We also set the numNodes and numOptionalNodes differently if any use of
$port_connected is present -- numNodes is the MINIMUM number of nodes
this device requires, and numOptionalNodes is how many additional
nodes may be specified.  numNodes+numOptionalNodes = what ADMS calls
"external" nodes.

There are *many* places throughout Xyce/ADMS where "what adms calls
external nodes" is used as a count of external nodes, and "what adms
calls internal nodes" is used as the count of internal nodes.  All of
those uses must be changed to take into account the fact that some
"external" nodes might not be because they're optional.  This commit
does NONE of that.  Miles to go before I sleep.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=390
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/24