Checkpoint work done on port_connected
This commit gets us a lot farther toward supporting port_connected in
simple cases.

Any use of $port_connected(node) in a Verilog-A model marks that node
as optional.  We do this both by sticking a flag in the module's copy
of the node object (which is used pretty much everywhere we loop over
nodes) and by adding the node to a list of optional nodes.

We keep a portsConnected_ vector of bools to reflect which nodes that
are designated as external as far as the module line is concerned are
actually used on the instance line --- any that are not are treated as
internal nodes.

This commit should correctly handle transitioning an optional node
between external or internal, which is mainly done just by changing
the number of external and internal nodes reported by the device to
topology.

No error checking is done (yet) to assure that all external nodes
after the first optional one are *also* optional.  This is an absolute
requirement for Xyce, but violating it does not yet make Xyce/ADMS
puke --- it'll gleefully and improperly generate bad code in such a
case (or at least, I think it will).

We take pains to handle lead currents correctly --- if a node is
optional and not given, we do NOT ask for a lead branch LID or store
its lead currents in the main lead current/voltage drop arrays.

At this point, NO work has been done to get node collapse done
correctly when it involves optional nodes.  In fact, I'm pretty sure
it'll all be wrong if it were tried with this version.

I have so far only been testing that a little toy problem with a
pointless $port_connected usage generated something that *looks* like
correct code, and that none of these changes have any impact at  all
on exisiting devices that don't use $port_connected.

There is a lot more work to do in order to get real devices that use
$port_connected working.  BSIMSOI is the real problem child, and it's
still out of reach.  This *might* be enough to make HiSIM work, but
there are probably still other issues to deal with.  I have not tried
yet, and won't be able to for a few days.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=390
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/24