Several small fixes related to error messages.
The functions that I created to replace "get_names", such as
getVoltageNodes, getLeadCurrents, etc, needed to "clear" the std::vector
of std::string that is passed in.   I had neglected to do this.  So, if
you called a bunch of these in a row, but passed in the same std::vector
to every one (b/c you were only doing this to check the presence of
certain things, not their details) then it would just keep adding to it.

This caused the error messages for bug 769 certification tests to be
slightly wrong.  In the first test, there is a .param that incorrectly
depends on a voltage node.  So, the error message is supposed to
complain about the voltage node.  However, with new expression, the
error message complained that there was a lead current (there was not).
Anyway, this is fixed now.

The other part of this commit is a new function for replacing parameter
names.  There is a narrow use case for this, for handling the tests from
bug 1801.    This is not a complete fix for this issue, but is the first
part of it.