Fix multiplicity terms in ifdeffed BSIM-SOI 4.6 code
When I added explicit multiplicity to the BSIM-SOI 4.6.1 model, I did
so neglecting the bits of code that are ifdeffed out by
PORT_CONNECTED (which at this time we don't use, because we don't
support $port_connected).

But I am working on getting $port_connected.  We will probably never
support it well enough to get an unmodified BSIM-SOI 4.x working (the
use of port_connected is insane here), but when I do get it working
well enough to support a HACKED BSIM-SOI, I need the terms to be
correct.  This fixes the few terms that were left out --- those that
set the thermal network when PORT_CONNECTED is defined.