Get port_connected working in a limited sense
This commit fixes a few mistakes in declaring of functions and
parameter passing, and a misuse of "min" instead of "std::min".

With this commit, Xyce/ADMS can process the BSIM6 with its
port_connected use into a plugin.  This is a low bar, as BSIM6's use
of port_connected is extremely simple and straightforward.

The code compiles and looks correct when I look at the source code and
what it does for the special cases.  However, I have not yet run any
tests of this plugin to see if it actually *works*.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=390
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/24