Add a few unit tests inspired by abmpow
This was prompted by some email discussions today.  The abmpow3 circuit
evaluates the power function for things like this:

pow(-2.5, 3.1)

If the inputs are real numbers (not complex) and the output is supposed
to be real, then this returns Nan.  The old expression library had some
error checking for this, but it apparently was not actually used.  And
the gold standard for this test also included Nan's.   So, that test was
a bit silly and is being updated.

This got me wondering about the new expression library.  It was not
failing the abmpow3 test.   But, unlike the old expression library, the
new expression library treats everything as complex numbers under the
covers, even if the calling code only wants a real numbered result.

I looked over the unit tests and didn't see any like the one above, in
which the base was negative and the exponent was not a whole number.
So, I've added one a test like that.  I've also added a few others based
on abmpow1.cir and abmpow2.cir.