Small error report fix for funcOp::output function
Each AST node used by the new expression library is an instance of one
of the many AST classes in the various ast files.    All of the AST
classes implement an "output" operator, which is what is used
recursively whenver the "dumpParseTree" function is called.

This commit makes a change to the output function of the
funcOp class, which handles function calls in an expression.  The funcOp
class has a functionNode_ pointer, which points to the top level AST
node of the function being called.   If the function hasn't been
resolved yet, then this node is null, or the RCP version of null.

The output function in the funcOp class checks to see if the
functionNode_ pointer is null.  If it is NOT null, then it calls
functionNode_->output().  Otherwise it doesn't make this call, because
this would cause a segfault.

But, it was then unconditionally calling the funcOp::val function, which
emits a fatal error when functionNode_ is null.

What this meant was that if I called exprPtr->dumpParseTree from
somewhere in the parser as a diagnostic, the code would exit with a
fatal error if the call had happened before the expression represented
by exprPtr had been fully resolved.  During parsing, this can easily
happen, and it was not a disirable effect.  That function is only ever
called as a diagnostic;  I don't want it to cause fatal errors.

Anyway, this is fixed now.