Re-establish "User Info" messages
Years ago, the error manager supported "info" messages as well as
"Error" or "Warning" --- the idea being that if a device or other
package simply needed to blat information at a user, "Info" was the
right class, since it's neither warning the user of something
unexpected or exiting with an error.

Though it was supported, it was rarely used except during development,
and when the error manager was refactored to provide various
stream-like objects instead of function calls, the capability was
retained under the hood but never exposed so it could be used.

Verilog-A models often use "$strobe()" to blat information at a user.
Most of these are actually warnings, or messages that precede
"$finish" (abort) calls.  But they're used fairly indiscriminately,
not only as warnings but sometimes just as informative (and possibly
unnecessary) noise.

My original implementation of $strobe simply blatted the text to
std::cerr.  This was suboptimal, not only because everything it
printed went to the error stream, but because the message printed
showed no context --- one could not in any way, shape, or form figure
out where this message was being produced or why.

I have decided to make use of the error handler package instead of
cerr, because devices can pass their instance object to the stream and
the resulting message will have the actual device instance that
produced the message printed as a prefix, complete with netlist file
name and line number.  But "UserWarning" is the wrong thing to use
here, because many of these $strobe uses aren't even warnings, they're
just noise.  Having them output with "Netlist warning in file...:"
would cornfuse users.

So I have added UserInfo and UserInfo0 functions to
ErrorHandlingPkg/N_ERH_Message.h file and the device
package's N_DEV_Message files.  These are used *exactly* like
UserWarning, but print "Netlist info in file ...:" instead.

The N_ERH_Message.C file already had code to handle "MSG_INFORMATION"
messages, but no mechanism had been provided to create such messages.