Make $strobe emit to UserInfo instead of cerr
Since I started Xyce/ADMS, I had made $strobe simply blat its
information directly to std::cerr.  This is not optimal, as the
messages get sent to the error stream with NO context at all.  If they
were important, one would have a hard time even knowing what device
created the message.

This commit changes $strobe to send information to UserInfo(*this)
instead of cerr.  In this way, every strobe output will now have "Netlist
info in file filename at or near line N:" and the actual device
instance name prefixing it.

This makes the output way more verbose, but actually associates it
with the user's input in a way that helps.

Most Verilog-A devices only use $strobe for warnings or error
messages.  A rare few (that is to say, EKV3) use it for huge blocks of
informative output selected with a model parameter.  For EKV3 with
that model parameter selected the output will be somewhat uglier than
before, as each line of output will be decorated with information
showing who output it.  But this is the rare case, and in the common
case this change will be a win for users.