Clean up "collapse to ground" for external nodes
Some Verilog-A models attempt to collapse external nodes to ground.
It is actually impossible to do this in Xyce --- once the netlist is
set up the nodes that appear on instance lines exist and are not
ground (unless, of course, they are actually 0 already).

However, it is not actually necessary for this to be done literally
--- a device can *pretend* the node is grounded simply by not loading
anything into its element of the F or Q vectors or their derivatives.

Doing this is almost certainly a path to singular matrices and
convergence failures.  But if the user actually has a well connected
circuit attached to the node that this one device is ignoring, there's
no problem.

This commit cleans up Xyce/ADMS to actually do collapse-to-ground of
external nodes this way.  In fact, in most aspects it was already
right, but this commit makes sure that we handle external lids
correctly, and that we NEVER actually collapse the jacobian stamp when
it's an external node being collapsed to ground (it's perfectly fine
to collapse an INTERNAL node to ground).  We do this by checking
portsConnected_ where necessary.

We also emit a LOUD and informative warning when this happens ---
since it is VERY likely to cause convergence problems, the user should
be told exactly what the problem is, that it could cause problems, and
how to fix it.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=390
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/24