Fix up collapse logic
Xyce/ADMS's logic for how to generate node collapse code was written
long before we tried to support optional nodes.

BSIM-SOI used enough node collapse associated with these optional
nodes that it fell into some holes that I'd forgotten to close up.

Notably, collapsing optional nodes onto something other than ground
was not handled correctly (it can work, but only if the node is not
connected) and there were some bits of code that had been written
assuming external nodes would never be collapsed that were wrong if
they *could* be --- for example, in loadNodeSymbols).

With these changes, I can get a version of BSIM-SOI 4.6 that has
optional nodes, and that works correctly when those optional nodes are
not given.  I do not yet know if it works when the optional nodes ARE
correctly specified.

https://joseki-vm.sandia.gov/bugzilla/show_bug?id=390
https://joseki-vm.sandia.gov/bugzilla/show_bug?id=1297
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/24
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/39