Add optional nodes to BSIM SOI 4.6.1
This commit hacks on the B4SOI 4.6.1 Verilog-A source code to disable
all questionable "collapse-to-ground" conditions of external nodes.
None of these can really work, and even when it *could* be possible to
let they slide (at the likely cost of having convergence problems),
many actually cause additional problems in Xyce/ADMS due to ambiguous
collapse issues (i.e. some nodes might be able to collapse onto two
different things, depending on model parameters).

This may be an inappropriate hack, and none of it has been tested yet.

The device does, however, work correctly when none of the optional
nodes are connected externally, so I am committing it as a checkpoint.

https://joseki-vm.sandia.gov/bugzilla/show_bug?id=1297
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/39