fix for bug 1801 regression tests
The bug 1801 regression tests are all about using global parameters in
the context of subcircuit parameters.    This particular use case is one
that I had not worked thru properly until now.

The bug 1801 "A1" circuit is this netlist:

.global_param resval=1

X1 1 0 RESSUB PARAMS: resistance={resval}
V1 1 0 5V

.step dec resval 1 10001 5
.dc v1 5 5 1
.print dc V(1) I(V1)

.subckt ressub a b params: resistance=100k
R1 a b {resistance}
.ends

So, in the X1 subcircuit instance, the value for the subcircuit argument
"resistance" is set to be equal to the global_param resval.
Additionally, "resval" is swept by a .STEP.

So, the expression for R1 includes the parameter resistances, so it has
to be attached to it.  And, "resistance" is ultimately a global_param
"resval" so it has to be attached to it.  I had hoped my machinery for
these attachments would be sufficient, but it was not.

Several changes were necessary to make this work.
(1) the "isConstant" machinery was busted, which I knew, but there
was still one place in the device entity that was attempting to
use it.  This broke the .STEP updates that should have changed
the value for "RESISTANCE"

(2) the param class in the new expression library assumed 2 kinds of
params: .param and .global_param.  I am not sure if this was really
necessary, but I added another category, for subcircuit params.  That
made some book keeping easier.

(3) Functions like newExpression::getNodes and getVariables, and
getParams all rely on certain internal data structures being set up in
the parser.  There is an AST tree, which is the primary storage location
for all the nodes.  However, there is also a bunch of STL containers for
differen types of objects.  So, there is a std::vector of voltage nodes,
a std::vector of current variabes, etc.  These are automatically
assembled during parsing, but they are incomplete whenever an external
expression is attached as a parameter or as a func.   I have a function
that corrects this, but that function wasn't being called enough, and so
for the above use case, all the "get" functions returned empty objects
and the IO parser assumed that the R1 expression was a constant and set
it to a hardwired value.    I experimented with several ways of fixing
this, and ultimately settled on one.

Anyway, with all of that done, the 1801 test now passes with the new
expression library.