Fix for bug 1806 certification test
The bug 1806 test is all about passing nodes into a subcircuit, which
has an expression inside of it that depends on those nodes.  To evaluate
the expression, the names of the nodes all have to be changed to the
ones being passed in.

There are a few gotchas in this, that the certification test was
intended to catch.  I hadn't handled those gotchas correctly.  The main
gotcha I didn't properly handle was when multiple nodes are renamed to
the same input node.  Like for example, if several of the input nodes
are set to ground.

This is fixed now.