Fix for bug 28 SON certification test.
The bug 28 SON test has a circuit node in it named "Int".

Earlier in my development of the expression library, I did not realize
that during parsing, many, many things that are clearly not intended to
be expressions are processed temporarily as expressions.  I haven't
checked if this always happens with circuit nodes.

Anyway, the result is that "Int" was then fed into the expression
library and parsed.  And, the expression library was processing it as a
syntax error, because it thought that "Int" should be the operator
Int(a,b).

I have updated the lexing/parsing of "Int" so that the token TOK_INT now
includes the left parenthesis.    That way, if you have "Int" by itself,
it will be given the more generic token, TOK_WORD, which can mean
practically anything, and will just be parsed as a yet-to-be-resolved
parameter.  That is still weird (IMO) but at least it won't exit with a
syntax error.

I have not bothered yet to do this with other operators.  In principal,
this probably means that (a) voltage nodes could, theoretically, have
names that match operator strings and (b) the expression library will
have to process them without emitting an error.  So, for many of the
operators in the expression library I should probably do this.

On one hand, I don't think many of the remaining broken regression
tests are busted due to this issue.  On the other hand, it only took me
about 5 minutes to do this for "Int", so I could probably do it for many
others pretty quickly.