Fix for bug 496 certification test.
The bug 496 test uses the "RAND()" function, but only to make sure that
the circuit runs.  It doesn't check the answer.  The circuit also
contains voltage node named RAND, which I am guessing was part of the
point of the test (to see if name conflicts were resolved/avoided)

The only reason the new expression library wasn't running this test is
that I had not implmented the randOp::val function  yet, and had it emit
a fatal error saying as much.

I have various plans for the random number related functions, including
this one.  But that's a bigger issue that I've been putting off for now.
For the most part, for the time being I'm having all of them just return
the values they should return if not doing sampling, which is generally
going to be the mean of the distribution.  (This is Hspice's behavior
when not doing sampling).  I had apparently not done this for randOp,
but I have now as of this commit.  It was a 2-line change.

For the purposes of this test, that is sufficient to get it to pass.