Fix for MULTIPLICITY/thermal_resistor test
I thought I had the issues between the thermal resistor and the new
expression library worked out, but I wasn't quite right about that.
All the thermal resistor tests except for this one have been passing.
I overlooked this one because the test didn't have a "thermalresistor"
tag.

The issue in this test is that (1) the thermal model is turned on for
the thermal resistor (2) the calculation is DC only, so the thermal
model *shouldn't* have any impact and (3) one of the .STEP commands
is sweeping over TEMP.  This combination of issues confused things.

The thermal resistor model really should have more logic inside of it
that is pertaining to transient vs. non-transient analysis.  The model
updates temperature based on (among other things)
the current time step size.  So, it really doesn't make any sense for
anything but transient.  But, there doesn't appear to be any machinery
turning these computations OFF for non-transient simulation.  That seems
weird to me.

The updated temperaturee isn't used for much, except for parameters
specificed as a temperature-dependent table.  In the DC case, these
temperature dependent tables should (IMO) just use the ambient
temperature, rather than the device temperature.  And that appears to
have been the case with the old expression library, although the code
was written in a way to completely obscure this.

This commit applies a minimal fix to the problem.  The model calls a
special version of "DeviceEntity::updateDependentParameters" which
passes in "temp" as a function argument.  I think this device is the
ONLY device that calls this version of that function.  All other devices
only call the argument free version of that function.

Since the thermal model is a transient model, it makes to sense to
update parameters using the results of that model for any non-transient
solve.  I had placed an if-statement around that function call that
tested "tranop", so that it wouldn't be called during the operating
point of a transient calculation.  But, that wasn't enough, as there are
DC tests that use this model.  I have now updated that if-statement to
also test for the dcopFlag.  That fixes the
MULTIPLICITY/thermal_resistor test.